Add Flask-Admin and Flask-WTF to reuirements.txt for form handling and an admin interface.
Add many more tables to the model and make the
first attempt at buildinjg an admin interface for
them.

Note: This is currently horribly insecure, there
is NO authentication or authorisation for the admin
interface. Do not put this anywhere where it can be
accessed over the internet while connected to a
real database.